refactor: split slide_list.rs below the 1000-line cap (F1-F10 growth)
The F1-F10 review fixes pushed slide_list.rs to 1145 lines, over the
quality gate's 1000-line hard cap (the Quality Checks CI failure on the
previous push). Pure mechanical extraction, zero behavior change:

- slide_save.rs now owns the whole per-slide save path: the moved
  save_all_fields_from_dom + get_field_value + capture_selection, plus
  the #556 helpers focused_slide_field_in_dom / restore_pending_focus /
  reconcile_after_seq_mismatch (all pub(super), same bodies).
- New slide_reorder.rs owns the drag-reorder drop handler
  (handle_slide_drop — the #552 failure badge + seq guard + F1
  order-merge + F4 typing flush/focus restore) and the song bubble
  renderer (render_song_bubble, F5's in-flow sticky row).
- slide_list.rs keeps the component itself: 1145 -> 836 prod lines
  (count_prod_lines.sh), comfortably under the cap.

Local gate re-run after the split: quality-check.sh --strict exit 0,
fmt + clippy (wasm, -D warnings) clean, presenter-ui cargo test --lib
235 passed, wasm-drag-drop.spec.ts 16 passed + wasm-slide-group-field
2 passed against the rebuilt binary.

Part of #556 (review response).
