+++ crates/presenter-ui/src/components/slide_list.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use leptos::prelude::*;
-use presenter_core::{resolve_sequence, Presentation, ResolvedSlide, Slide};
+use presenter_core::{resolve_sequence, ResolvedSlide};
 use wasm_bindgen::JsCast;
 
 use crate::api;
@@ -12,242 +12,12 @@ use crate::utils::color::group_pill_style;
 use super::slide_list_scroll::{handle_wheel_event, scroll_slide_into_view, scroll_slides_to_top};
 use super::slide_list_utils::{
     apply_focused_class, field_has_warning, format_multiline, is_interactive_tag,
-    reorder_slide_ids, resolve_group_display, slide_has_any_warning,
+    resolve_group_display, slide_has_any_warning,
 };
+use super::slide_reorder::{handle_slide_drop, render_song_bubble};
 use super::slide_save::{
-    finish_save_status_err, finish_save_status_ok, save_with_status, start_save_status,
+    capture_selection, reconcile_after_seq_mismatch, save_all_fields_from_dom,
 };
-
-/// Get a textarea/input value from the DOM by slide_id and field name.
-fn get_field_value(doc: &web_sys::Document, slide_id: &str, field: &str) -> String {
-    let selector = format!(
-        "[data-slide-id=\"{}\"] [data-field=\"{}\"]",
-        slide_id, field
-    );
-    if let Ok(Some(el)) = doc.query_selector(&selector) {
-        if let Ok(ta) = el.clone().dyn_into::<web_sys::HtmlTextAreaElement>() {
-            return ta.value();
-        }
-        if let Ok(inp) = el.dyn_into::<web_sys::HtmlInputElement>() {
-            return inp.value();
-        }
-    }
-    String::new()
-}
-
-/// Unified save function that saves ALL fields from DOM atomically.
-/// This prevents data loss when editing multiple fields before blur.
-/// Takes `RwSignal` (which is `Copy`) instead of `&AppContext` to allow
-/// use in multiple `move` closures without moving the entire context.
-fn save_all_fields_from_dom(
-    pres_id: &str,
-    slide_id: &str,
-    _current_field: &str,
-    _sel_start: u32,
-    _sel_end: u32,
-    selected_pres: RwSignal<Option<presenter_core::Presentation>>,
-    op: &OperatorState,
-) {
-    let doc = crate::utils::window::document();
-
-    let main = get_field_value(&doc, slide_id, "main");
-    let translation = get_field_value(&doc, slide_id, "translation");
-    let stage = get_field_value(&doc, slide_id, "stage");
-    let group_val = get_field_value(&doc, slide_id, "group");
-    let group = if group_val.trim().is_empty() {
-        None
-    } else {
-        Some(group_val.trim().to_string())
-    };
-
-    // Skip no-op saves.
-    let pres = selected_pres.get_untracked();
-    if let Some(p) = &pres {
-        if let Some(slide) = p.slides.iter().find(|s| s.id.to_string() == slide_id) {
-            let orig = &slide.content;
-            let orig_group = orig.group.as_ref().map(|g| g.name().to_string());
-            if orig.main.value() == main
-                && orig.translation.value() == translation
-                && orig.stage.value() == stage
-                && orig_group == group
-            {
-                return;
-            }
-        }
-    }
-
-    // #515: bump the edit generation SYNCHRONOUSLY, the moment a genuine
-    // (non-no-op) edit is committed to saving — NOT when the save's network
-    // round-trip later completes. A `get_presentation` refetch (opening /
-    // re-opening the presentation) captures this counter before it fires;
-    // bumping here, before the async save even starts, guarantees that ANY
-    // such refetch still in flight at this point is provably stale no
-    // matter which of the two async calls happens to resolve first.
-    op.slide_edit_seq.update(|seq| *seq += 1);
-
-    save_with_status(
-        pres_id.to_string(),
-        slide_id.to_string(),
-        main,
-        translation,
-        stage,
-        group,
-        op.save_status,
-        selected_pres,
-        op.groups_version,
-    );
-}
-
-/// The slide id + field name currently holding focus, per an actual DOM
-/// check (`document.activeElement`) — not `OperatorState`'s
-/// `focused_slide_id`/`focused_field`, which only record the LAST field
-/// that was focused and are never cleared on blur. Used by the reorder
-/// drop handler (#556 F4) to know whether there is genuinely in-flight
-/// typing to protect before a tracked `selected_pres` update tears down
-/// and rebuilds the (unkeyed) slide list.
-fn focused_slide_field_in_dom() -> Option<(String, String)> {
-    let doc = crate::utils::window::document();
-    let active = doc.active_element()?;
-    let field = active.get_attribute("data-field")?;
-    let card = active.closest("[data-slide-id]").ok().flatten()?;
-    let slide_id = card.get_attribute("data-slide-id")?;
-    Some((slide_id, field))
-}
-
-/// Re-focus the given slide's field after a tracked slide-list rebuild
-/// (#556 F4). Best-effort: if the slide (or the field) no longer exists
-/// post-rebuild — e.g. it was concurrently deleted — this is a silent
-/// no-op, there is nothing left to focus.
-fn restore_pending_focus(slide_id: &str, field: &str) {
-    let doc = crate::utils::window::document();
-    let selector = format!("[data-slide-id=\"{slide_id}\"] [data-field=\"{field}\"]");
-    if let Ok(Some(el)) = doc.query_selector(&selector) {
-        if let Ok(html_el) = el.dyn_into::<web_sys::HtmlElement>() {
-            let _ = html_el.focus();
-        }
-    }
-}
-
-/// Reconcile an editing response (reorder / insert / duplicate / delete)
-/// against a possible `slide_edit_seq` mismatch (#556 F1/F3): some OTHER
-/// edit landed while this request was in flight. Never apply a response
-/// outright once it's known-stale, but never silently keep showing
-/// "Saved ✓" over a client that displays structure/content the server
-/// doesn't have either.
-///
-/// - If the response's slide-id SET still matches what's currently
-///   displayed, the concurrent edit only changed CONTENT (e.g. a
-///   content-field save bumped the seq) — apply the server's ORDER while
-///   keeping the (newer) local CONTENT for each slide.
-/// - If the id sets differ, a structural change (insert/duplicate/delete)
-///   raced this call and the merge above can't be done meaningfully — issue
-///   ONE guarded refetch and apply it only if nothing else has landed by
-///   the time it resolves.
-async fn reconcile_after_seq_mismatch(
-    pres_id: String,
-    server_slides: Vec<Slide>,
-    selected_pres: RwSignal<Option<Presentation>>,
-    slide_edit_seq: RwSignal<u64>,
-) {
-    let local_ids: Option<Vec<String>> = selected_pres
-        .get_untracked()
-        .map(|p| p.slides.iter().map(|s| s.id.to_string()).collect());
-    let server_ids: Vec<String> = server_slides.iter().map(|s| s.id.to_string()).collect();
-
-    let same_id_set = local_ids
-        .map(|mut local| {
-            let mut server_sorted = server_ids.clone();
-            local.sort();
-            server_sorted.sort();
-            local == server_sorted
-        })
-        .unwrap_or(false);
-
-    if same_id_set {
-        selected_pres.update(|p| {
-            if let Some(pres) = p.as_mut() {
-                let mut by_id: HashMap<String, Slide> = pres
-                    .slides
-                    .drain(..)
-                    .map(|s| (s.id.to_string(), s))
-                    .collect();
-                pres.slides = server_ids
-                    .iter()
-                    .filter_map(|id| by_id.remove(id))
-                    .collect();
-            }
-        });
-        return;
-    }
-
-    // A structural change raced this request: one guarded refetch, applied
-    // only if nothing newer has landed by the time it resolves.
-    let seq_at_refetch = slide_edit_seq.get_untracked();
-    if let Ok(detail) = api::presentations::get_presentation(&pres_id).await {
-        if slide_edit_seq.get_untracked() == seq_at_refetch {
-            selected_pres.set(Some(detail.presentation));
-        }
-    }
-}
-
-/// Capture current selection range from a textarea event.
-fn capture_selection(ev: &web_sys::Event) -> (u32, u32) {
-    if let Some(target) = ev.target() {
-        if let Ok(ta) = target.clone().dyn_into::<web_sys::HtmlTextAreaElement>() {
-            let start = ta.selection_start().ok().flatten().unwrap_or(0);
-            let end = ta.selection_end().ok().flatten().unwrap_or(0);
-            return (start, end);
-        }
-        if let Ok(inp) = target.dyn_into::<web_sys::HtmlInputElement>() {
-            let start = inp.selection_start().ok().flatten().unwrap_or(0);
-            let end = inp.selection_end().ok().flatten().unwrap_or(0);
-            return (start, end);
-        }
-    }
-    (0, 0)
-}
-
-/// Renders the draggable song bubble above the slide list.
-/// Returns a reactive closure that shows the bubble when a presentation is
-/// selected, or nothing otherwise.
-fn render_song_bubble(ctx: AppContext, op: OperatorState) -> impl IntoView {
-    move || {
-        let presentation = ctx.selected_presentation.get();
-        let Some(pres) = presentation else {
-            return "".into_any();
-        };
-        let pres_id = pres.id.to_string();
-        let pres_id_drag = pres_id.clone();
-        let pres_name = pres.name.clone();
-        let op_drag = op.clone();
-        let op_end = op.clone();
-        view! {
-            <div
-                class="operator__slides-bubble"
-                data-role="slides-song-bubble"
-                data-presentation-id=pres_id
-                draggable="true"
-                title="Drag into a playlist"
-                on:dragstart=move |ev: web_sys::DragEvent| {
-                    if let Some(dt) = ev.data_transfer() {
-                        let _ = dt.set_data("text/plain", &pres_id_drag);
-                        let _ = dt.set_data("application/x-presentation-id", &pres_id_drag);
-                        dt.set_effect_allowed("copy");
-                    }
-                    op_drag.search_dragging.set(true);
-                    op_drag.dragging_from_search.set(true);
-                }
-                on:dragend=move |_| {
-                    op_end.search_dragging.set(false);
-                    op_end.dragging_from_search.set(false);
-                }
-            >
-                <span class="operator__slides-bubble-name">{pres_name}</span>
-            </div>
-        }
-        .into_any()
-    }
-}
 
 #[component]
 pub fn SlideList() -> impl IntoView {
@@ -390,91 +160,12 @@ pub fn SlideList() -> impl IntoView {
                                 ev.prevent_default();
                             }
                         }
+                        // #552/#556 F1/F4: the whole guarded reorder round-trip
+                        // (failure badge, slide_edit_seq staleness guard,
+                        // order-merge reconciliation, in-flight-typing flush +
+                        // focus restore) lives in `slide_reorder::handle_slide_drop`.
                         on:drop=move |ev: web_sys::DragEvent| {
-                            ev.prevent_default();
-                            if let Some(dragged_id) = op_drop.dragging_slide_id.get_untracked() {
-                                if let Some(target) = ev.target() {
-                                    if let Ok(el) = target.dyn_into::<web_sys::Element>() {
-                                        if let Some(card) = el.closest("[data-slide-id]").ok().flatten() {
-                                            let target_id = card.get_attribute("data-slide-id").unwrap_or_default();
-                                            if let Some(p) = ctx.selected_presentation.get_untracked() {
-                                                let pres_id = p.id.to_string();
-                                                let ids: Vec<String> = p.slides.iter().map(|s| s.id.to_string()).collect();
-                                                if let Some(new_ids) = reorder_slide_ids(ids, &dragged_id, &target_id) {
-                                                    let selected_pres = ctx.selected_presentation;
-                                                    // #552: a reorder used to swallow any network/server error
-                                                    // silently (no `else` branch at all) — a WiFi blip on the
-                                                    // tablet made the drop look exactly like "nothing happened",
-                                                    // with no visible failure and no way to tell the drag from a
-                                                    // no-op. Reuse the same per-slide save-status badge the
-                                                    // content-edit fields already show, keyed on the dragged
-                                                    // slide, so a failed reorder is visible instead of silent.
-                                                    //
-                                                    // Also bump `slide_edit_seq` the same way a content edit does
-                                                    // (slide_edit_seq.update() above the async call, checked after
-                                                    // it resolves) so a slower, overlapping EARLIER drag's response
-                                                    // can't silently clobber a newer one that already landed.
-
-                                                    // #556 F4: a reorder's own tracked apply below MUST rebuild
-                                                    // the (unkeyed) slide list — the new ORDER changes which
-                                                    // cards render where, unlike a content save. Flush whatever
-                                                    // the user is mid-typing BEFORE that rebuild (both now, at
-                                                    // drop time, and again right before the response is applied
-                                                    // below) so nothing typed is silently lost, and restore
-                                                    // focus to the same field once the DOM exists again.
-                                                    if let Some((focus_sid, focus_field)) = focused_slide_field_in_dom() {
-                                                        save_all_fields_from_dom(&pres_id, &focus_sid, &focus_field, 0, 0, selected_pres, &op_drop);
-                                                    }
-
-                                                    op_drop.slide_edit_seq.update(|seq| *seq += 1);
-                                                    let my_seq = op_drop.slide_edit_seq.get_untracked();
-                                                    let slide_edit_seq = op_drop.slide_edit_seq;
-                                                    let save_status = op_drop.save_status;
-                                                    let op_for_apply = op_drop.clone();
-                                                    let token = start_save_status(&dragged_id, save_status);
-                                                    leptos::task::spawn_local(async move {
-                                                        match api::presentations::reorder_slides(&pres_id, new_ids).await {
-                                                            Ok(slides) => {
-                                                                // #556 F1: on a seq mismatch, never silently
-                                                                // discard the response while still showing
-                                                                // "Saved ✓" over a stale client order/content.
-                                                                //
-                                                                // #556 F4: flush + restore focus around whichever
-                                                                // apply path runs below (direct or reconciled) —
-                                                                // BOTH are tracked updates that rebuild the DOM.
-                                                                let focused = focused_slide_field_in_dom();
-                                                                if let Some((ref fid, ref ffield)) = focused {
-                                                                    save_all_fields_from_dom(&pres_id, fid, ffield, 0, 0, selected_pres, &op_for_apply);
-                                                                }
-                                                                if slide_edit_seq.get_untracked() == my_seq {
-                                                                    selected_pres.update(|p| {
-                                                                        if let Some(pres) = p.as_mut() {
-                                                                            pres.slides = slides;
-                                                                        }
-                                                                    });
-                                                                } else {
-                                                                    reconcile_after_seq_mismatch(pres_id, slides, selected_pres, slide_edit_seq).await;
-                                                                }
-                                                                if let Some((fid, ffield)) = focused {
-                                                                    gloo_timers::callback::Timeout::new(0, move || {
-                                                                        restore_pending_focus(&fid, &ffield);
-                                                                    })
-                                                                    .forget();
-                                                                }
-                                                                finish_save_status_ok(dragged_id, save_status, token).await;
-                                                            }
-                                                            Err(_) => {
-                                                                finish_save_status_err(dragged_id, save_status, token);
-                                                            }
-                                                        }
-                                                    });
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            op_drop.dragging_slide_id.set(None);
+                            handle_slide_drop(ev, ctx.selected_presentation, &op_drop);
                         }
                     >
                     // #556 F5: the song bubble is now an IN-FLOW, sticky first
